Mostly working
I ran the program and the output seems to be mostly correct. There are
two situations that I tested that are producing the wrong output. I have
commented them in the Definitions.ss file. I Also made a few more
functions to test f-l. I am pretty sure that the bug is minor in the
program.