Changes to SchemeParser
Made changes to the SchemeParser, Attempting to create new scope based
on the current token and push that new scope to the symbol table stack
and then add the elements to that scope…

Still needs work:
Still have to give the symbolTable a reference to the root node of the
scope, and the root node needs a reference to the symbolTable of its
scope. I am not exactly sure hot to implement this.